Use double instead of float cast of sample count in transport internal sync update.

+++ src/engine/transport.cpp
@@ -291,7 +291,7 @@ void Transport::_update_internals()
 void Transport::_update_internal_sync(int64_t samples)
 {
     // We cannot assume chunk size is an absolute multiple of samples for all buffer sizes.
-    auto chunks_passed = static_cast<float>(samples) / AUDIO_CHUNK_SIZE;
+    auto chunks_passed = static_cast<double>(samples) / AUDIO_CHUNK_SIZE;
 
     if (_playmode != _set_playmode)
     {
@@ -301,7 +301,7 @@ void Transport::_update_internal_sync(int64_t samples)
         _rt_event_dispatcher->send_event(RtEvent::make_playing_mode_event(0, _set_playmode));
     }
 
-    double beats_per_chunk =  _set_tempo / 60.0 * static_cast<double>(AUDIO_CHUNK_SIZE) / _samplerate;
+    double beats_per_chunk = _set_tempo / 60.0 * static_cast<double>(AUDIO_CHUNK_SIZE) / _samplerate;
     _beats_per_chunk = beats_per_chunk;
 
     if (_state_change == PlayStateChange::STARTING) // Reset bar beat count when starting
